refactor: break service cycles and improve stability
