+++ App/utils/TokenManager.ts
@@ -1,5 +1,6 @@
 import { getDocument, setDocument } from '@/services/firestoreService';
 import { ensureAuth } from '@/utils/authGuard';
+import { getIdToken as fbGetIdToken, getCurrentUserId as fbGetCurrentUserId } from '@/lib/auth';
 
 export const getTokenCount = async () => {
   const uid = await ensureAuth();
@@ -62,6 +63,29 @@ export const syncSubscriptionStatus = async () => {
     await setDocument(`users/${uid}`, { tokens: 9999 });
   }
 };
+
+export async function getIdToken(forceRefresh = false) {
+  return fbGetIdToken(forceRefresh);
+}
+
+export async function getCurrentUserId(): Promise<string | null> {
+  return fbGetCurrentUserId();
+}
+
+export async function getAuthHeader() {
+  const token = await getIdToken(true);
+  if (!token) throw new Error('Unable to refresh ID token');
+  return { Authorization: `Bearer ${token}` };
+}
+
+export async function getAuthHeaders() {
+  const { Authorization } = await getAuthHeader();
+  return { Authorization, 'Content-Type': 'application/json' };
+}
+
+export async function getToken(forceRefresh = false) {
+  return getIdToken(forceRefresh);
+}
 
 export function init() {
   console.log('✅ TokenManager initialized');
